Move useQuestion hook in seperate file
